Okay making real progress again. The menu  is printing, there are entries with differences it found within the first five entries so this will be a process.
The next step is setting up the system for handling menu interaction.

This will involve adding another row to the table, the 'edit' row, and then to the right of row an arrow showing which row is currently seleted, if one is. Once a row has been selected, the user will selecte B, P, or N to set if the selected entry is correct in BP, is correct in PN, or if neither is correct. Then the command line will print "Mark row (number) [name] as (Bp/Pn/Neither) being the corret choice: y/n." the user will hit y/n and then be sent back to the first menu. This selection will dictate which list the entry gets placed into, for saving and sharing.

The user selects their command after hte table is printed, by entering one fo several options: h will show the help menu listing all commands. e will start the edit option, to change one of the choices, f to finish editing the entry. When entering the choice the console will print: "Command: " and wait for the letter entry. The command shoudl also be able to be typed as the full text for the command.