Removed the name check for the match system. Put it behind a flag so it can easily be used if one desires.
Next is figuring out the spacing issue on text from BP.

Once that is done, I have to clean up and confirm the file path stuff, and then its just going over and making things a bit better.

Then I'll need to update the readme to explain the directory structure needed to run the program.