Working to get the CR stuff mentioned in #7 functional. The currnet version will crash the program so I do not recommend using it.
But now when the user selects BP as correct for a CR entry in the data-matcher, that entry gets put into a special list. Later, when the program is processing the various lists, it checks this list.

Then it goes through each entry and pulls the CR from each BP entry. Then it pulls the relevant reviws from the matching XML entry.

Then it compares the list of CR entries to the list of review enties. If there are any matches, those matching entries get set aside.

now with a list of entries which are not in the review ptrs, we check the whole list of XMl entries. If any match the year, we then check if any entries in that year match pages. If we do, we assume one of those matches, if there is one or more, is the correct match.

Any entries which do not match, the program then generates a new entry type, the CRREviewData.cs type. This class basically just has the data from the CR entry in it.

The only thing missing right now, besides testing, is searching the excel file (which I need to convert to a csv) and then compare the journal names extracted from the CR entry, and find the relevant jounral ID number, to fill in the appearsIn ptr set.

Once that is done, in theory at least, this newest update should be finished.